+++ odoo/local-src/specific_stock/models/stock_picking.py
@@ -4,7 +4,7 @@
 
 from datetime import date
 
-from odoo import models, api, _
+from odoo import fields, models, api, _
 from odoo.exceptions import Warning
 from odoo.tools import DEFAULT_SERVER_DATE_FORMAT as DATE_FORMAT
 from odoo.exceptions import UserError
@@ -101,3 +101,15 @@ class StockPicking(models.Model):
             })
 
         return result
+
+
+class StockReturnPicking(models.TransientModel):
+    _inherit = "stock.return.picking"
+
+    @api.model
+    def default_get(self, fields):
+        """ Set by default to refund """
+        res = super(StockReturnPicking, self).default_get(fields)
+        for line in res.get('product_return_moves', {}):
+            line[2]['to_refund_so'] = True
+        return res