+++ odoo/local-src/specific_stock/__manifest__.py
@@ -52,6 +52,7 @@
         'wizards/stock_receive_lot.xml',
         'wizards/put_in_pack_helper.xml',
         'wizards/generate_voice_identifier.xml',
+        'wizards/open_uop.xml',
 
         # Views (loaded after wizards)
         'views/stock_picking.xml',