+++ odoo/local-src/specific_print/models/__init__.py
@@ -3,3 +3,4 @@ from . import stock
 from . import printer
 from . import res_partner
 from . import res_company
+from . import product_template