+++ odoo/local-src/specific_print/__manifest__.py
@@ -26,6 +26,7 @@
         'report/stock_lot_label.xml',
         'views/stock.xml',
         'views/product_template.xml',
+        'views/stock_production_lot.xml',
         ],
     'installable': True,
     'license': 'AGPL-3',