+++ HISTORY.rst
@@ -8,7 +8,11 @@ Unreleased
 
 **Features and Improvements**
 
+* ALCYN-1669: Mark customer return as to be refunded by default
+
 **Bugfixes**
+
+* ALCYN-1602: Never set a scrap location on a reception pack operation
 
 **Build**