+++ HISTORY.rst
@@ -21,6 +21,7 @@ Unreleased
 
 * ALCYN-1602: Never set a scrap location on a reception pack operation
 * ALCYN-1670: Fix cancel remaining quantity to deliver on a sales order
+* ALCYN-1487: Fix sale exception translations
 
 **Build**