+++ HISTORY.rst
@@ -11,6 +11,7 @@ Unreleased
 * ALCYN-1623: Sort bins on products to take first bins for label printing
 * ALCYN-1683: Speed up stock export by using direct sql for the sale_average computation (ESB)
 * ALCYN-1685: Disable location translation
+* Update setup of locations for "Aliments". Define rows for "reserve" and set picking_zone
 
 **Bugfixes**