+++ HISTORY.rst
@@ -10,6 +10,8 @@ Unreleased
 
 **Bugfixes**
 
+* ALCYN-1649: sale_triple_discount: Fix tax total on report
+
 **Build**
 
 **Documentation**