+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 **Features and Improvements**
 
 * ALCYN-1667: Allow to search a product by the reference of the supplier
+* ALCYN-1684: Force the user to put in pack before validate a picking
 
 **Bugfixes**