+++ HISTORY.rst
@@ -10,6 +10,8 @@ Unreleased
 
 **Bugfixes**
 
+* ALCYN-1686: Use the date done on the delivery slip and on the exported CSV
+
 **Build**
 
 **Documentation**